Two small improvements that I've made to the C implementations while recoding dv_parse_ac_coeffs_pass0. [1] dv_decode_vlc() now expects its bits parameter to be left-justified to bit 15.  This means that the caller can just pull the next 16 bits in the stream regardless of whether it's going to call dv_decode_vlc() or __dv_decode_vlc(). [2] reorder indices are now gint8 instead of gint.  Smaller is better, and means less fuss for pointer arithmetic.

+++ dv.h
@@ -54,8 +54,8 @@ typedef struct dv_block_s {
   double       fcoeffs[64];
   gint         dct_mode; 
   gint         class_no;
-  gint        *reorder;
-  gint        *reorder_sentinel;
+  gint8        *reorder;
+  gint8        *reorder_sentinel;
   gint         offset;   // bitstream offset of first unused bit 
   gint         end;      // bitstream offset of last bit + 1
   gint         eob;
